# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: KWG Directory API
  version: 0.1.0
  description: OpenChain Korea Work Group 멤버 주소록 REST API. 인증은 next-auth 세션 쿠키 기반. /api/auth/* 는 next-auth 내부 핸들러로 본 문서 범위 밖.
  license:
    name: Apache-2.0
servers:
  - url: /
    description: 동일 오리진
...
paths:
...
  /api/members/search:
    get:
      summary: 멤버 검색·필터·정렬(페이지네이션).
      tags:
        - members
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 100
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - name
              - recent
              - random
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 검색 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberSearchResult'
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/upload/avatar:
    post:
      summary: 아바타 이미지 업로드(multipart). 2MB·JPEG/PNG/WebP.
      tags:
        - upload
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 업로드됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUploadResult'
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        name_ko:
          type: string
        name_en:
          type: string
          nullable: true
        company:
          type: string
        role:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        category:
          type: string
          enum:
            - 기업
            - 연구/공공
            - 학계
          nullable: true
        email:
          type: string
          nullable: true
          description: 비로그인 시 redact
        email_public:
          type: boolean
        linkedin:
          type: string
          nullable: true
        github:
          type: string
          nullable: true
        discord:
          type: string
          nullable: true
        blog:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        approved:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name_ko
        - company
        - approved
    MemberList:
      type: array
      items:
        $ref: '#/components/schemas/Member'
...
    MemberSearchResult:
      type: object
      properties:
        members:
          $ref: '#/components/schemas/MemberList'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required:
        - members
        - total
        - page
        - pageSize
        - totalPages
...
    AvatarUploadResult:
      type: object
      properties:
        url:
          type: string
          format: uri
      required:
        - url
...